requirements.txt
pillow-heif>=0.12.0
pillow>=10.0.0
